+++ app/assets/javascripts/folio/input/collection_remote_select.js
@@ -1,7 +1,51 @@
+//= require folio/i18n
+
 window.Folio = window.Folio || {}
 window.Folio.Input = window.Folio.Input || {}
 
 window.Folio.Input.CollectionRemoteSelect = {}
+window.Folio.Input.CollectionRemoteSelect.AUTOCOMPLETE_QUERY_MIN_LENGTH = 3
+window.Folio.Input.CollectionRemoteSelect.i18n = {
+  cs: {
+    inputTooShort: `Zadejte alespoň ${window.Folio.Input.CollectionRemoteSelect.AUTOCOMPLETE_QUERY_MIN_LENGTH} znaky.`
+  },
+  en: {
+    inputTooShort: `Type at least ${window.Folio.Input.CollectionRemoteSelect.AUTOCOMPLETE_QUERY_MIN_LENGTH} characters.`
+  }
+}
+
+window.Folio.Input.CollectionRemoteSelect.BlankOrMinimumInputLength = function (decorated, $element, options) {
+  this.minimumInputLength = options.get('minimumInputLength')
+
+  decorated.call(this, $element, options)
+}
+
+window.Folio.Input.CollectionRemoteSelect.BlankOrMinimumInputLength.prototype.query = function (decorated, params, callback) {
+  params.term = params.term || ''
+
+  if (params.term.length > 0 && params.term.length < this.minimumInputLength) {
+    this.trigger('results:message', {
+      message: 'inputTooShort',
+      args: {
+        minimum: this.minimumInputLength,
+        input: params.term,
+        params
+      }
+    })
+
+    return
+  }
+
+  decorated.call(this, params, callback)
+}
+
+window.Folio.Input.CollectionRemoteSelect.dataAdapter = () => {
+  const amd = window.jQuery.fn.select2.amd
+  const Utils = amd.require('select2/utils')
+  const AjaxData = amd.require('select2/data/ajax')
+
+  return Utils.Decorate(AjaxData, window.Folio.Input.CollectionRemoteSelect.BlankOrMinimumInputLength)
+}
 
 window.Folio.Input.CollectionRemoteSelect.setValue = (input, value) => {
   const $input = window.jQuery(input)
@@ -17,7 +61,14 @@ window.Folio.Input.CollectionRemoteSelect.bind = (input, { includeBlank, url }) 
 
   $input.select2({
     width: '100%',
-    language: document.documentElement.lang,
+    language: [
+      {
+        inputTooShort: () => window.Folio.i18n(window.Folio.Input.CollectionRemoteSelect.i18n, 'inputTooShort')
+      },
+      document.documentElement.lang
+    ],
+    minimumInputLength: window.Folio.Input.CollectionRemoteSelect.AUTOCOMPLETE_QUERY_MIN_LENGTH,
+    dataAdapter: window.Folio.Input.CollectionRemoteSelect.dataAdapter(),
     allowClear: true,
     placeholder: { id: '', text: includeBlank },
     dropdownCssClass: $input.data('dropdown-class') || '',
@@ -25,7 +76,6 @@ window.Folio.Input.CollectionRemoteSelect.bind = (input, { includeBlank, url }) 
       url: url || $input.data('url'),
       dataType: 'JSON',
       delay: 250,
-      minimumInputLength: 0,
       cache: false,
       data: (params) => {
         const data = {